Deploys GitBook documentation to GitHub Pages
Sets up a GitHub Actions workflow to automatically build and deploy the KTON documentation to GitHub Pages.

This allows for easier maintenance and ensures the documentation is always up-to-date with the latest changes. It improves the accessibility of KTON documentation.

Also, refactors documentation by embedding images locally and adding a Telegram Mini App link.